pyproject.toml
[tool.mypy]
exclude = [
    'venv',
    '.venv',
    '.github',
    'env',
    '.env',
    'tests',
]
plugins = [
    "sqlalchemy.ext.mypy.plugin",
    "pydantic.mypy",
]
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode="auto"
addopts = "--strict-markers --tb=short -s -v -p no:cacheprovider"

[tool.coverage.run]
omit = [
    "src/tests/*",
    "src/core/*",
    "src/db/*",
    "src/schemas/*",
    "src/migrations/*",
    "__init__.py",
]

[tool.poetry]
name = "async-pythn-sprint-5"
version = "0.1.0"
description = ""
authors = ["Nikita Fomin <fominquant@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.2"}
asyncpg = "^0.28.0"
psycopg2-binary = "^2.9.9"
databases = {extras = ["postgresql"], version = "^0.8.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
aerich = "^0.7.2"
sqlalchemy = "^1.4"
alembic = "^1.12.0"
greenlet = "^3.0.0"
aiofiles = "^23.2.1"
py7zr = "^0.20.6"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
flake8 = "^6.1.0"
mypy = "^1.6.0"
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
